+++ src/main/java/com/google/firebase/cloud/FirestoreClient.java
@@ -11,6 +11,8 @@ import com.google.firebase.FirebaseApp;
 import com.google.firebase.ImplFirebaseTrampolines;
 import com.google.firebase.internal.FirebaseService;
 import com.google.firebase.internal.NonNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * {@code FirestoreClient} provides access to Google Cloud Firestore. Use this API to obtain a
@@ -25,6 +27,8 @@ import com.google.firebase.internal.NonNull;
  * methods, this API will throw a runtime exception.
  */
 public class FirestoreClient {
+
+  private static final Logger logger = LoggerFactory.getLogger(FirestoreClient.class);
 
   private final Firestore firestore;
 
@@ -48,7 +52,9 @@ public class FirestoreClient {
   }
 
   /**
-   * Returns the Firestore instance associated with the default Firebase app.
+   * Returns the Firestore instance associated with the default Firebase app. Returns the same
+   * instance for all invocations. The Firestore instance and all references obtained from it
+   * becomes unusable, once the default app is deleted.
    *
    * @return A non-null <a href="https://googlecloudplatform.github.io/google-cloud-java/google-cloud-clients/apidocs/com/google/cloud/firestore/Firestore.html">{@code Firestore}</a>
    *     instance.
@@ -59,7 +65,9 @@ public class FirestoreClient {
   }
 
   /**
-   * Returns the Firestore instance associated with the specified Firebase app.
+   * Returns the Firestore instance associated with the specified Firebase app. For a given app,
+   * always returns the same instance. The Firestore instance and all references obtained from it
+   * becomes unusable, once the specified app is deleted.
    *
    * @param app A non-null {@link FirebaseApp}.
    * @return A non-null <a href="https://googlecloudplatform.github.io/google-cloud-java/google-cloud-clients/apidocs/com/google/cloud/firestore/Firestore.html">{@code Firestore}</a>
@@ -89,10 +97,11 @@ public class FirestoreClient {
 
     @Override
     public void destroy() {
-      // NOTE: We don't explicitly tear down anything here (for now). User won't be able to call
-      // FirestoreClient.getFirestore() any more, but already created Firestore instances will
-      // continue to work. Request Firestore team to provide a cleanup/teardown method on the
-      // Firestore object.
+      try {
+        instance.firestore.close();
+      } catch (Exception e) {
+        logger.warn("Error while closing the Firestore instance", e);
+      }
     }
   }